error handling for missing Exception Notifier config

+++ config/environments/production.rb
@@ -3,7 +3,8 @@ Meegoqa::Application.configure do
 
   # Configure exception notifications
   exception_notifier_config_file_path = File.expand_path('../../exception_notifier', __FILE__)
-  email_addresses = eval File.open(exception_notifier_config_file_path).gets
+  email_addresses = eval File.open(exception_notifier_config_file_path).gets if File.exist?(exception_notifier_config_file_path)
+
   config.middleware.use "::ExceptionNotifier",
     :email_prefix => "[MeeGo QA Reports] ",
     :sender_address => %{"Exception Notifier" <notifier@qa-reports.meego.com>},
